+++ src/constants.js
@@ -3,6 +3,7 @@ export const itemStatuses = {
   ON_ORDER: 'On order',
   AVAILABLE: 'Available',
   IN_TRANSIT: 'In transit',
+  IN_PROCESS: 'In process',
   AWAITING_PICKUP: 'Awaiting pickup',
   PAGED: 'Paged',
 };