+++ test/bigtest/interactors/inventory.js
@@ -5,7 +5,14 @@ import {
   selectable,
   fillable,
   clickable,
+  text
 } from '@bigtest/interactor';
+
+@interactor class SearchFilter {
+  static defaultScope = '#input-inventory-search-qindex';
+
+  all = text('option[value=all]');
+}
 
 export default @interactor class InventoryInteractor {
   static defaultScope = '[data-test-inventory-instances]';
@@ -15,6 +22,8 @@ export default @interactor class InventoryInteractor {
   instance = scoped('[data-test-instance-details]');
 
   chooseFilter = selectable('#input-inventory-search-qindex');
+  filter = new SearchFilter();
+
   fillFilter = fillable('#input-inventory-search');
   clickSearch = clickable('[data-test-search-and-sort-submit]');
 }