+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Retrieve 1000 items from lookup tables. Refs UIIN-413.
 * Fix alphabetical order of lookup tables. Fixes UIIN-421.
+* Don't append wildcards to search terms. Fixes UIIN-481.
 
 ## [1.7.0](https://github.com/folio-org/ui-inventory/tree/v1.7.0) (2019-03-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.6.0...v1.7.0)