+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * Don't append wildcards to search terms. Fixes UIIN-481.
 * Provide the correct a11y props to `<Layer>`s.
 * Add missing order-by clauses to item-record lookup tables. Fixes UIIN-522.
+* Send `contributors.type.primary` as correct type. Refs MODINV-117.
 
 ## [1.7.0](https://github.com/folio-org/ui-inventory/tree/v1.7.0) (2019-03-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.6.0...v1.7.0)