+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 * Move acquisitions items to the acquisitions accordion. Fixes UIIN-506.
 * Remove duplicate asterisks from required fields. Fixes UIIN-490.
 * Add ability to mark Paged items as missing. Part of UIIN-555.
+* Mark In process items as missing. Part of UIIN-560.
 
 ## [1.7.0](https://github.com/folio-org/ui-inventory/tree/v1.7.0) (2019-03-15)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v1.6.0...v1.7.0)